move the QuickPickHoverDelegate into quickinput platform
This involves standardizing `IHoverWidget` coming from `iconHoverDelegate`
