+++ src/vs/editor/standalone/browser/quickInput/standaloneQuickInputService.ts
@@ -19,6 +19,8 @@ import { QuickInputController, IQuickInputControllerHost } from 'vs/platform/qui
 import { QuickInputService } from 'vs/platform/quickinput/browser/quickInputService';
 import { createSingleCallFunction } from 'vs/base/common/functional';
 import { IQuickAccessController } from 'vs/platform/quickinput/common/quickAccess';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { IHoverService } from 'vs/platform/hover/browser/hover';
 
 class EditorScopedQuickInputService extends QuickInputService {
 
@@ -29,9 +31,18 @@ class EditorScopedQuickInputService extends QuickInputService {
 		@IInstantiationService instantiationService: IInstantiationService,
 		@IContextKeyService contextKeyService: IContextKeyService,
 		@IThemeService themeService: IThemeService,
-		@ICodeEditorService codeEditorService: ICodeEditorService
+		@ICodeEditorService codeEditorService: ICodeEditorService,
+		@IConfigurationService configurationService: IConfigurationService,
+		@IHoverService hoverService: IHoverService,
 	) {
-		super(instantiationService, contextKeyService, themeService, new EditorScopedLayoutService(editor.getContainerDomNode(), codeEditorService));
+		super(
+			instantiationService,
+			contextKeyService,
+			themeService,
+			new EditorScopedLayoutService(editor.getContainerDomNode(), codeEditorService),
+			configurationService,
+			hoverService
+		);
 
 		// Use the passed in code editor as host for the quick input widget
 		const contribution = QuickInputEditorContribution.get(editor);
