+++ src/vs/platform/quickinput/browser/quickInput.ts
@@ -8,7 +8,7 @@ import { StandardKeyboardEvent } from 'vs/base/browser/keyboardEvent';
 import { ActionBar } from 'vs/base/browser/ui/actionbar/actionbar';
 import { Button, IButtonStyles } from 'vs/base/browser/ui/button/button';
 import { CountBadge, ICountBadgeStyles } from 'vs/base/browser/ui/countBadge/countBadge';
-import { IHoverDelegate } from 'vs/base/browser/ui/iconLabel/iconHoverDelegate';
+import { IHoverDelegate, IHoverDelegateOptions, IHoverWidget } from 'vs/base/browser/ui/iconLabel/iconHoverDelegate';
 import { IInputBoxStyles } from 'vs/base/browser/ui/inputbox/inputBox';
 import { IKeybindingLabelStyles } from 'vs/base/browser/ui/keybindingLabel/keybindingLabel';
 import { IListRenderer, IListVirtualDelegate } from 'vs/base/browser/ui/list/list';
@@ -31,6 +31,8 @@ import { IInputBox, IKeyMods, IQuickInput, IQuickInputButton, IQuickInputHideEve
 import { QuickInputBox } from './quickInputBox';
 import { QuickInputList, QuickInputListFocus } from './quickInputList';
 import { getIconClass, renderQuickInputDescription } from './quickInputUtils';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { IHoverService } from 'vs/platform/hover/browser/hover';
 
 export interface IQuickInputOptions {
 	idPrefix: string;
@@ -47,7 +49,11 @@ export interface IQuickInputOptions {
 		renderers: IListRenderer<T, any>[],
 		options: IListOptions<T>,
 	): List<T>;
-	hoverDelegate?: IHoverDelegate;
+	/**
+	 * @todo With IHover in vs/editor, can we depend on the service directly
+	 * instead of passing it through a hover delegate?
+	 */
+	hoverDelegate: IHoverDelegate;
 	styles: IQuickInputStyles;
 }
 
@@ -1250,5 +1256,48 @@ export class QuickWidget extends QuickInput implements IQuickWidget {
 
 		this.ui.setVisibilities(visibilities);
 		super.update();
+	}
+}
+
+export class QuickInputHoverDelegate implements IHoverDelegate {
+	private lastHoverHideTime = 0;
+	readonly placement = 'element';
+
+	get delay() {
+		if (Date.now() - this.lastHoverHideTime < 200) {
+			return 0; // show instantly when a hover was recently shown
+		}
+
+		return this.configurationService.getValue<number>('workbench.hover.delay');
+	}
+
+	constructor(
+		private readonly configurationService: IConfigurationService,
+		private readonly hoverService: IHoverService
+	) { }
+
+	showHover(options: IHoverDelegateOptions, focus?: boolean): IHoverWidget | undefined {
+		// Only show the hover hint if the content is of a decent size
+		const showHoverHint = (
+			options.content instanceof HTMLElement
+				? options.content.textContent ?? ''
+				: typeof options.content === 'string'
+					? options.content
+					: options.content.value
+		).length > 20;
+		return this.hoverService.showHover({
+			...options,
+			persistence: {
+				hideOnKeyDown: false,
+			},
+			appearance: {
+				showHoverHint,
+				skipFadeInAnimation: true,
+			},
+		}, focus);
+	}
+
+	onDidHideHover(): void {
+		this.lastHoverHideTime = Date.now();
 	}
 }
