+++ src/vs/platform/quickinput/browser/quickInputService.ts
@@ -18,8 +18,10 @@ import { IInputBox, IInputOptions, IKeyMods, IPickOptions, IQuickInputButton, IQ
 import { defaultButtonStyles, defaultCountBadgeStyles, defaultInputBoxStyles, defaultKeybindingLabelStyles, defaultProgressBarStyles, defaultToggleStyles, getListStyles } from 'vs/platform/theme/browser/defaultStyles';
 import { activeContrastBorder, asCssVariable, pickerGroupBorder, pickerGroupForeground, quickInputBackground, quickInputForeground, quickInputListFocusBackground, quickInputListFocusForeground, quickInputListFocusIconForeground, quickInputTitleBackground, widgetBorder, widgetShadow } from 'vs/platform/theme/common/colorRegistry';
 import { IThemeService, Themable } from 'vs/platform/theme/common/themeService';
-import { IQuickInputOptions, IQuickInputStyles } from './quickInput';
+import { IQuickInputOptions, IQuickInputStyles, QuickInputHoverDelegate } from './quickInput';
 import { QuickInputController, IQuickInputControllerHost } from 'vs/platform/quickinput/browser/quickInputController';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { IHoverService } from 'vs/platform/hover/browser/hover';
 
 export class QuickInputService extends Themable implements IQuickInputService {
 
@@ -59,7 +61,9 @@ export class QuickInputService extends Themable implements IQuickInputService {
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 		@IContextKeyService protected readonly contextKeyService: IContextKeyService,
 		@IThemeService themeService: IThemeService,
-		@ILayoutService protected readonly layoutService: ILayoutService
+		@ILayoutService protected readonly layoutService: ILayoutService,
+		@IConfigurationService protected readonly configurationService: IConfigurationService,
+		@IHoverService private readonly hoverService: IHoverService
 	) {
 		super(themeService);
 	}
@@ -86,7 +90,8 @@ export class QuickInputService extends Themable implements IQuickInputService {
 				renderers: IListRenderer<T, any>[],
 				options: IWorkbenchListOptions<T>
 			) => this.instantiationService.createInstance(WorkbenchList, user, container, delegate, renderers, options) as List<T>,
-			styles: this.computeStyles()
+			styles: this.computeStyles(),
+			hoverDelegate: new QuickInputHoverDelegate(this.configurationService, this.hoverService)
 		};
 
 		const controller = this._register(new QuickInputController({
