+++ src/vs/workbench/contrib/preferences/browser/settingsEditorSettingIndicators.ts
@@ -21,7 +21,8 @@ import { IUserDataSyncEnablementService } from 'vs/platform/userDataSync/common/
 import { SettingsTreeSettingElement } from 'vs/workbench/contrib/preferences/browser/settingsTreeModels';
 import { POLICY_SETTING_TAG } from 'vs/workbench/contrib/preferences/common/preferences';
 import { IWorkbenchConfigurationService } from 'vs/workbench/services/configuration/common/configuration';
-import { IHoverOptions, IHoverService, IHoverWidget } from 'vs/platform/hover/browser/hover';
+import { IHoverOptions, IHoverService } from 'vs/platform/hover/browser/hover';
+import { IHoverWidget } from 'vs/base/browser/ui/iconLabel/iconHoverDelegate';
 
 const $ = DOM.$;
